src/backend/pyproject.toml
[project]
name = "kratos-agent-service"
version = "0.1.0"
description = "Agentic AI backend service powered by GitHub Copilot SDK & Microsoft Foundry"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "azure-identity>=1.17.0",
    "azure-cosmos>=4.7.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-search-documents>=11.6.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b26",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
    "github-copilot-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
